+++ src/entities/message/DiceType.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace tsvetkov\telegram_bot\entities\message;
+
+/**
+ * Class DiceType
+ * @package tsvetkov\telegram_bot\entities\message
+ */
+abstract class DiceType
+{
+    const DICE = '🎲';
+    const TARGET = '🎯';
+    const BASKETBALL = '🏀';
+}